+++ manager/cloud/gcp/storage.ts
